Rest of controllers, services etc

+++ topicsapp/src/main/java/com/siwz/app/services/implementation/SubjectManager.java
@@ -21,8 +21,8 @@ public class SubjectManager implements SubjectService {
 
     private final TopicService topicService;
 
-    @Override // TAG subjects
-    public Long createSubject(Subject subject) throws ApplicationException { // TODO api endpoint POST /subjects
+    @Override
+    public Long createSubject(Subject subject) throws ApplicationException {
         if(subjectRepository.existsByName(subject.getName())) {
             throw new ApplicationException(DAOError.DAO_SUBJECT_ALREADY_EXISTS, subject.getName());
         }
@@ -30,8 +30,8 @@ public class SubjectManager implements SubjectService {
         return subject.getId();
     }
 
-    @Override // TAG subjects
-    public void updateSubject(Long subjectId, Subject newSubject) throws ApplicationException { // TODO api endpoint PATCH /subjects/{id}
+    @Override
+    public void updateSubject(Long subjectId, Subject newSubject) throws ApplicationException {
         Optional<Subject> originalSubject = subjectRepository.findById(subjectId);
         if(!originalSubject.isPresent()) {
             throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
@@ -39,42 +39,27 @@ public class SubjectManager implements SubjectService {
         updateSubjectData(originalSubject.get(), newSubject);
     }
 
-    @Override // TAG subjects
-    public void deleteSubject(Long subjectId) { // TODO api endpoint DELETE /subjects/{id}
+    @Override
+    public void deleteSubject(Long subjectId) throws ApplicationException {
+        if(!subjectRepository.existsById(subjectId)) {
+            throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
+        }
         subjectRepository.deleteById(subjectId);
     }
 
-    @Override // TAG subjects
-    public List<Subject> getAllSubjects() { // TODO api endpoint GET /subjects
+    @Override
+    public List<Subject> getAllSubjects() {
         return subjectRepository.findAll();
     }
 
-    @Override // TAG subjects
-    public Subject getSubjectById(Long subjectId) throws ApplicationException { // TODO api endpoint GET /subjects/{id}
+    @Override
+    public Subject getSubjectById(Long subjectId) throws ApplicationException {
         return subjectRepository.findById(subjectId)
                 .orElseThrow(() -> new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId));
     }
 
-    @Override // TAG subject-topics
-    public List<Topic> getSubjectTopics(Long subjectId) throws ApplicationException { // TODO api endpoint GET /subjects/{id}/topics
-        Optional<Subject> subject = subjectRepository.findById(subjectId);
-        if(!subject.isPresent()) {
-            throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
-        }
-        return topicService.getTopics(subject.get());
-    }
-
-    @Override // TAG subject-topics
-    public Topic getSubjectTopicById(Long subjectId, Long topicId) throws ApplicationException { // TODO api endpoint GET /subjects/{id}/topics/{id}
-        Optional<Subject> subject = subjectRepository.findById(subjectId);
-        if(!subject.isPresent()) {
-            throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
-        }
-        return topicService.getTopicById(topicId, subject.get());
-    }
-
-    @Override // TAG subject-topics
-    public Long createSubjectTopic(Long subjectId, Topic topic) throws ApplicationException { // TODO api endpoint POST /subjects/{id}/topics
+    @Override
+    public Long createSubjectTopic(Long subjectId, Topic topic) throws ApplicationException {
         Optional<Subject> subject = subjectRepository.findById(subjectId);
         if(!subject.isPresent()) {
             throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
@@ -83,8 +68,8 @@ public class SubjectManager implements SubjectService {
         return topicService.createTopic(topic);
     }
 
-    @Override // TAG subject-topics
-    public void updateSubjectTopic(Long subjectId, Long topicId, Topic topic) throws ApplicationException { // TODO api endpoint PATCH /subjects/{id}/topics/{id}
+    @Override
+    public void updateSubjectTopic(Long subjectId, Long topicId, Topic topic) throws ApplicationException {
         Optional<Subject> subject = subjectRepository.findById(subjectId);
         if(!subject.isPresent()) {
             throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
@@ -93,13 +78,31 @@ public class SubjectManager implements SubjectService {
         topicService.updateTopic(topicId, topic);
     }
 
-    @Override //TAG subject-topics
-    public void deleteSubjectTopic(Long subjectId, Long topicId) throws ApplicationException { // TODO api endpoint DELETE /subjects/{id}/topics/{id}
+    @Override
+    public void deleteSubjectTopic(Long subjectId, Long topicId) throws ApplicationException {
         Optional<Subject> subject = subjectRepository.findById(subjectId);
         if(!subject.isPresent()) {
             throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
         }
         topicService.deleteTopic(topicId, subject.get());
+    }
+
+    @Override
+    public List<Topic> getSubjectTopics(Long subjectId) throws ApplicationException {
+        Optional<Subject> subject = subjectRepository.findById(subjectId);
+        if(!subject.isPresent()) {
+            throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
+        }
+        return topicService.getTopics(subject.get());
+    }
+
+    @Override
+    public Topic getSubjectTopicById(Long subjectId, Long topicId) throws ApplicationException {
+        Optional<Subject> subject = subjectRepository.findById(subjectId);
+        if(!subject.isPresent()) {
+            throw new ApplicationException(DAOError.DAO_SUBJECT_NOT_FOUND, subjectId);
+        }
+        return topicService.getTopicById(topicId, subject.get());
     }
 
     @Override
